requirements.txt
amqp==2.5.2
Babel==2.7.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
celery==4.3.0
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==2.2.7
django-bootstrap4==1.0.1
django-csvimport==2.13
django-debug-toolbar==2.1
django-mathfilters==0.4.0
django-pwa==1.0.5
django-registration==3.0.1
django-tz-detect==0.2.10
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
flower==0.9.3
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
pkg-resources==0.0.0
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==5.1.1
vine==1.3.0
zipp==0.6.0
